requirements.txt
mss==4.0.2
Pillow==5.4.1
